+++ cubi_sak/exceptions.py
@@ -17,8 +17,12 @@ class IrodsIcommandsUnavailableException(CubiSakException):
     """Raised when iRODS icommands are not available but required."""
 
 
-class UnsupportedIsaTabFeaturException(CubiSakException):
+class UnsupportedIsaTabFeatureException(CubiSakException):
     """Raised when an unsupported ISA-tab feature occurs."""
+
+
+class InvalidIsaTabException(CubiSakException):
+    """Raised when ISA-tab breaks an assumption."""
 
 
 class MissingFileException(CubiSakException):