+++ tests/test_snappy_itransfer_ngs_mapping.py
@@ -1,10 +1,11 @@
-"""Tests for ``cubi_sak.snappy.itransfer_raw_data``.
+"""Tests for ``cubi_sak.snappy.itransfer_ngs_mapping``.
 
 We only run some smoke tests here.
 """
 
 import os
 from unittest import mock
+from unittest.mock import ANY
 
 import pytest
 from pyfakefs import fake_filesystem
@@ -12,10 +13,10 @@ from pyfakefs import fake_filesystem
 from cubi_sak.__main__ import setup_argparse, main
 
 
-def test_run_snappy_itransfer_raw_data_help(capsys):
+def test_run_snappy_itransfer_ngs_mapping_help(capsys):
     parser, subparsers = setup_argparse()
     with pytest.raises(SystemExit) as e:
-        parser.parse_args(["snappy", "itransfer-raw-data", "--help"])
+        parser.parse_args(["snappy", "itransfer-ngs-mapping", "--help"])
 
     assert e.value.code == 0
 
@@ -24,11 +25,11 @@ def test_run_snappy_itransfer_raw_data_help(capsys):
     assert not res.err
 
 
-def test_run_snappy_itransfer_raw_data_nothing(capsys):
+def test_run_snappy_itransfer_ngs_mapping_nothing(capsys):
     parser, subparsers = setup_argparse()
 
     with pytest.raises(SystemExit) as e:
-        parser.parse_args(["snappy", "itransfer-raw-data"])
+        parser.parse_args(["snappy", "itransfer-ngs-mapping"])
 
     assert e.value.code == 2
 
@@ -37,13 +38,14 @@ def test_run_snappy_itransfer_raw_data_nothing(capsys):
     assert res.err
 
 
-def test_run_snappy_itransfer_raw_data_smoke_test(mocker):
+def test_run_snappy_itransfer_ngs_mapping_smoke_test(mocker):
     fake_base_path = "/base/path"
     dest_path = "/irods/dest"
     tsv_path = os.path.join(os.path.dirname(__file__), "data", "germline.out")
     argv = [
+        "--verbose",
         "snappy",
-        "itransfer-raw-data",
+        "itransfer-ngs-mapping",
         "--num-parallel-transfers",
         "1",
         "--base-path",
@@ -60,18 +62,32 @@ def test_run_snappy_itransfer_raw_data_smoke_test(mocker):
     for member in ("index", "father", "mother"):
         for ext in ("", ".md5"):
             fake_file_paths.append(
-                "%s/ngs_mapping/work/input_links/%s-N1-DNA1-WES1/%s-N1-DNA1-WES1.fastq.gz%s"
+                "%s/ngs_mapping/work/bwa.%s-N1-DNA1-WES1/out/%s-N1-DNA1-WES1.fastq.gz%s"
+                % (fake_base_path, member, member, ext)
+            )
+            fs.create_file(fake_file_paths[-1])
+            fake_file_paths.append(
+                "%s/ngs_mapping/work/bwa.%s-N1-DNA1-WES1/log/bwa.%s-N1-DNA1-WES1.log%s"
                 % (fake_base_path, member, member, ext)
             )
             fs.create_file(fake_file_paths[-1])
 
+    # Remove index's log MD5 file again so it is recreated.
+    fs.remove(fake_file_paths[3])
+
     fake_os = fake_filesystem.FakeOsModule(fs)
     mocker.patch("glob.os", fake_os)
     mocker.patch("cubi_sak.snappy.itransfer_common.os", fake_os)
-    mocker.patch("cubi_sak.snappy.itransfer_raw_data.os", fake_os)
+    mocker.patch("cubi_sak.snappy.itransfer_ngs_mapping.os", fake_os)
 
     mock_check_output = mock.mock_open()
     mocker.patch("cubi_sak.snappy.itransfer_common.check_output", mock_check_output)
+
+    fake_open = fake_filesystem.FakeFileOpen(fs)
+    mocker.patch("cubi_sak.snappy.itransfer_common.open", fake_open)
+
+    mock_check_call = mock.mock_open()
+    mocker.patch("cubi_sak.snappy.itransfer_common.check_call", mock_check_call)
 
     # Actually exercise code and perform test.
     parser, subparsers = setup_argparse()
@@ -79,14 +95,26 @@ def test_run_snappy_itransfer_raw_data_smoke_test(mocker):
     res = main(argv)
 
     assert not res
+
     # We do not care about call order but simply test call count and then assert that all files are there which would
     # be equivalent of comparing sets of files.
+
+    assert fs.exists(fake_file_paths[3])
+
+    assert mock_check_call.call_count == 1
+    mock_check_call.assert_called_once_with(
+        ["md5sum", "bwa.index-N1-DNA1-WES1.log"],
+        cwd=os.path.dirname(fake_file_paths[3]),
+        stdout=ANY,
+    )
+
     assert mock_check_output.call_count == len(fake_file_paths) * 2
     for path in fake_file_paths:
-        index, rel_path = os.path.relpath(
-            path, os.path.join(fake_base_path, "ngs_mapping/work/input_links")
+        mapper_index, rel_path = os.path.relpath(
+            path, os.path.join(fake_base_path, "ngs_mapping/work")
         ).split("/", 1)
-        remote_path = os.path.join(dest_path, index, "raw_data", args.remote_dir_date, rel_path)
+        _mapper, index = mapper_index.split(".", 1)
+        remote_path = os.path.join(dest_path, index, "ngs_mapping", args.remote_dir_date, rel_path)
         expected_mkdir_argv = ["imkdir", "-p", os.path.dirname(remote_path)]
         expected_irsync_argv = ["irsync", "-a", "-K", path, "i:%s" % remote_path]
         mock_check_output.assert_any_call(expected_mkdir_argv)