+++ HISTORY.md
@@ -8,6 +8,8 @@
 - Adding `cubi-sak snappy pull-sheet`.
 - Converting `snappy-transfer_utils`, adding `cubi-sak snappy ...`
     - `itransfer-raw-data`
+    - `itransfer-ngs-mapping`
+- Adding `mypy` checks to CI.
 
 ## v0.1.0